backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
numpy==2.2.3
pandas==2.2.3
scikit-learn==1.6.1
torch==2.6.0
joblib==1.4.2
pytest==8.3.4
python-dateutil==2.9.0.post0
requests==2.32.3
